rename layouts to layout
